refactor: breakup ofm.ts into multiple modules
extracts a collection of modules containing plugins and helpers to make
the obsidian flavored markdown plugin more digestable and easier to
manage.
