+++ quartz/plugins/transformers/ofm.ts
@@ -1,23 +1,25 @@
 import { QuartzTransformerPlugin } from "../types"
-import { Root, Html, BlockContent, DefinitionContent, Paragraph, Code } from "mdast"
-import { Element, Literal, Root as HtmlRoot } from "hast"
+import { Root } from "mdast"
+import { Element, Root as HtmlRoot } from "hast"
 import { ReplaceFunction, findAndReplace as mdastFindReplace } from "mdast-util-find-and-replace"
-import { slug as slugAnchor } from "github-slugger"
 import rehypeRaw from "rehype-raw"
-import { SKIP, visit } from "unist-util-visit"
-import path from "path"
 import { JSResource } from "../../util/resources"
-// @ts-ignore
-import calloutScript from "../../components/scripts/callout.inline.ts"
-// @ts-ignore
-import checkboxScript from "../../components/scripts/checkbox.inline.ts"
-import { FilePath, pathToRoot, slugTag, slugifyFilePath } from "../../util/path"
-import { toHast } from "mdast-util-to-hast"
-import { toHtml } from "hast-util-to-html"
-import { PhrasingContent } from "mdast-util-find-and-replace/lib"
-import { capitalize } from "../../util/lang"
+import { pathToRoot } from "../../util/path"
 import { PluggableList } from "unified"
-import { ListItem } from "mdast-util-to-hast/lib/handlers/list-item"
+
+import { arrowRegexReplacement } from "./obsidian/arrows"
+import { blockReferencePlugin } from "./obsidian/blockReferences"
+import { calloutPlugin, calloutResources, calloutTextTransform } from './obsidian/callouts';
+import { checkboxPlugin, checkboxResources } from "./obsidian/checkbox"
+import { commentsTextTransform } from './obsidian/comments';
+import { markdownHTMLEmbedReplacements } from "./obsidian/embedHTML"
+import { videoEmbedPlugin } from "./obsidian/embedVideo"
+import { embedYoutubePlugin } from "./obsidian/embedYoutube"
+import { highlightRegexReplacement } from "./obsidian/highlights"
+import { mermaidPlugin, mermaidResources } from "./obsidian/mermaid"
+import { tagsRegexReplacement } from "./obsidian/tags"
+import { customTaskStatusPlugin } from "./obsidian/tasks"
+import { wikilinkRegexReplacment, wikilinkTextTransform } from "./obsidian/wikilinks"
 
 export interface Options {
   comments: boolean
@@ -32,6 +34,7 @@ export interface Options {
   enableYouTubeEmbed: boolean
   enableVideoEmbed: boolean
   enableCheckbox: boolean
+  customTaskStatuses: boolean
 }
 
 const defaultOptions: Options = {
@@ -47,146 +50,24 @@ const defaultOptions: Options = {
   enableYouTubeEmbed: true,
   enableVideoEmbed: true,
   enableCheckbox: false,
+  customTaskStatuses: true,
 }
-
-const calloutMapping = {
-  note: "note",
-  abstract: "abstract",
-  summary: "abstract",
-  tldr: "abstract",
-  info: "info",
-  todo: "todo",
-  tip: "tip",
-  hint: "tip",
-  important: "tip",
-  success: "success",
-  check: "success",
-  done: "success",
-  question: "question",
-  help: "question",
-  faq: "question",
-  warning: "warning",
-  attention: "warning",
-  caution: "warning",
-  failure: "failure",
-  missing: "failure",
-  fail: "failure",
-  danger: "danger",
-  error: "danger",
-  bug: "bug",
-  example: "example",
-  quote: "quote",
-  cite: "quote",
-} as const
-
-const arrowMapping: Record<string, string> = {
-  "->": "&rarr;",
-  "-->": "&rArr;",
-  "=>": "&rArr;",
-  "==>": "&rArr;",
-  "<-": "&larr;",
-  "<--": "&lArr;",
-  "<=": "&lArr;",
-  "<==": "&lArr;",
-}
-
-function canonicalizeCallout(calloutName: string): keyof typeof calloutMapping {
-  const normalizedCallout = calloutName.toLowerCase() as keyof typeof calloutMapping
-  // if callout is not recognized, make it a custom one
-  return calloutMapping[normalizedCallout] ?? calloutName
-}
-
-export const externalLinkRegex = /^https?:\/\//i
-
-export const arrowRegex = /(-{1,2}>|={1,2}>|<-{1,2}|<={1,2})/g
-
-// !?                -> optional embedding
-// \[\[              -> open brace
-// ([^\[\]\|\#]+)    -> one or more non-special characters ([,],|, or #) (name)
-// (#[^\[\]\|\#]+)?  -> # then one or more non-special characters (heading link)
-// (\|[^\[\]\#]+)?   -> | then one or more non-special characters (alias)
-export const wikilinkRegex = /!?\[\[([^\[\]\|\#]+)?(#+[^\[\]\|\#]+)?(\|[^\[\]\#]+)?\]\]/g
-const highlightRegex = /==([^=]+)==/g
-const commentRegex = /%%[\s\S]*?%%/g
-// from https://github.com/escwxyz/remark-obsidian-callout/blob/main/src/index.ts
-const calloutRegex = /^\[\!(\w+)\]([+-]?)/
-const calloutLineRegex = /^> *\[\!\w+\][+-]?.*$/gm
-
-// (?:^| )                   -> non-capturing group, tag should start be separated by a space or be the start of the line
-// #(...)                    -> capturing group, tag itself must start with #
-// (?:[-_\p{L}\d\p{Z}])+     -> non-capturing group, non-empty string of (Unicode-aware) alpha-numeric characters and symbols, hyphens and/or underscores
-// (?:\/[-_\p{L}\d\p{Z}]+)*) -> non-capturing group, matches an arbitrary number of tag strings separated by "/"
-const tagRegex = /(?:^| )#((?:[-_\p{L}\p{Emoji}\p{M}\d])+(?:\/[-_\p{L}\p{Emoji}\p{M}\d]+)*)/gu
-const blockReferenceRegex = /\^([-_A-Za-z0-9]+)$/g
-const ytLinkRegex = /^.*(youtu.be\/|v\/|u\/\w\/|embed\/|watch\?v=|\&v=)([^#\&\?]*).*/
-const ytPlaylistLinkRegex = /[?&]list=([^#?&]*)/
-const videoExtensionRegex = /\.(mp4|webm|ogg|avi|mov|flv|wmv|mkv|mpg|mpeg|3gp|m4v)$/
-const wikilinkImageEmbedRegex = /^(?<alt>(?!^\d*x?\d*$).*?)?(\|?\s*?(?<width>\d+)(x(?<height>\d+))?)?$/
-const extendedTaskItemRegex = /^\[([^\[\]])\] (.*)/gm;
 
 export const ObsidianFlavoredMarkdown: QuartzTransformerPlugin<Partial<Options> | undefined> = (
   userOpts,
 ) => {
   const opts = { ...defaultOptions, ...userOpts }
 
-  const mdastToHtml = (ast: PhrasingContent | Paragraph) => {
-    const hast = toHast(ast, { allowDangerousHtml: true })!
-    return toHtml(hast, { allowDangerousHtml: true })
-  }
-
   return {
     name: "ObsidianFlavoredMarkdown",
-    textTransform(_ctx, src) {
-      // do comments at text level
-      if (opts.comments) {
-        if (src instanceof Buffer) {
-          src = src.toString()
-        }
-
-        src = src.replace(commentRegex, "")
-      }
-
-      // pre-transform blockquotes
-      if (opts.callouts) {
-        if (src instanceof Buffer) {
-          src = src.toString()
-        }
-
-        src = src.replace(calloutLineRegex, (value) => {
-          // force newline after title of callout
-          return value + "\n> "
-        })
-      }
-
-      // pre-transform wikilinks (fix anchors to things that may contain illegal syntax e.g. codeblocks, latex)
-      if (opts.wikilinks) {
-        if (src instanceof Buffer) {
-          src = src.toString()
-        }
-
-        src = src.replace(wikilinkRegex, (value, ...capture) => {
-          const [rawFp, rawHeader, rawAlias]: (string | undefined)[] = capture
-
-          const fp = rawFp ?? ""
-          const anchor = rawHeader?.trim().replace(/^#+/, "")
-          const blockRef = Boolean(anchor?.startsWith("^")) ? "^" : ""
-          const displayAnchor = anchor ? `#${blockRef}${slugAnchor(anchor)}` : ""
-          const displayAlias = rawAlias ?? rawHeader?.replace("#", "|") ?? ""
-          const embedDisplay = value.startsWith("!") ? "!" : ""
-
-          if (rawFp?.match(externalLinkRegex)) {
-            return `${embedDisplay}[${displayAlias.replace(/^\|/, "")}](${rawFp})`
-          }
-
-          return `${embedDisplay}[[${fp}${displayAnchor}${displayAlias}]]`
-        })
-      }
-
+    textTransform(_ctx, src: string | Buffer) {
+      if (opts.comments) commentsTextTransform(src);
+      if (opts.callouts) calloutTextTransform(src);
+      if (opts.wikilinks) wikilinkTextTransform(src);
       return src
     },
     markdownPlugins(ctx) {
       const plugins: PluggableList = []
-      const cfg = ctx.cfg.configuration
 
       // regex replacements
       plugins.push(() => {
@@ -194,332 +75,20 @@ export const ObsidianFlavoredMarkdown: QuartzTransformerPlugin<Partial<Options> 
           const replacements: [RegExp, string | ReplaceFunction][] = []
           const base = pathToRoot(file.data.slug!)
 
-          if (opts.wikilinks) {
-            replacements.push([
-              wikilinkRegex,
-              (value: string, ...capture: string[]) => {
-                let [rawFp, rawHeader, rawAlias] = capture
-                const fp = rawFp?.trim() ?? ""
-                const anchor = rawHeader?.trim() ?? ""
-                const alias = rawAlias?.slice(1).trim()
+          if (opts.wikilinks) replacements.push(wikilinkRegexReplacment)
+          if (opts.highlight) replacements.push(highlightRegexReplacement)
+          if (opts.parseArrows) replacements.push(arrowRegexReplacement)
+          if (opts.parseTags) replacements.push(tagsRegexReplacement(file, base))
+          if (opts.enableInHtmlEmbed) markdownHTMLEmbedReplacements(tree, replacements)
 
-                // embed cases
-                if (value.startsWith("!")) {
-                  const ext: string = path.extname(fp).toLowerCase()
-                  const url = slugifyFilePath(fp as FilePath)
-                  if ([".png", ".jpg", ".jpeg", ".gif", ".bmp", ".svg", ".webp"].includes(ext)) {
-                    const match = wikilinkImageEmbedRegex.exec(alias ?? "")
-                    const alt = match?.groups?.alt ?? ""
-                    const width = match?.groups?.width ?? "auto"
-                    const height = match?.groups?.height ?? "auto"
-                    return {
-                      type: "image",
-                      url,
-                      data: {
-                        hProperties: {
-                          width,
-                          height,
-                          alt,
-                        },
-                      },
-                    }
-                  } else if ([".mp4", ".webm", ".ogv", ".mov", ".mkv"].includes(ext)) {
-                    return {
-                      type: "html",
-                      value: `<video src="${url}" controls></video>`,
-                    }
-                  } else if (
-                    [".mp3", ".webm", ".wav", ".m4a", ".ogg", ".3gp", ".flac"].includes(ext)
-                  ) {
-                    return {
-                      type: "html",
-                      value: `<audio src="${url}" controls></audio>`,
-                    }
-                  } else if ([".pdf"].includes(ext)) {
-                    return {
-                      type: "html",
-                      value: `<iframe src="${url}"></iframe>`,
-                    }
-                  } else {
-                    const block = anchor
-                    return {
-                      type: "html",
-                      data: { hProperties: { transclude: true } },
-                      value: `<blockquote class="transclude" data-url="${url}" data-block="${block}"><a href="${
-                        url + anchor
-                      }" class="transclude-inner">Transclude of ${url}${block}</a></blockquote>`,
-                    }
-                  }
-
-                  // otherwise, fall through to regular link
-                }
-
-                // internal link
-                const url = fp + anchor
-                return {
-                  type: "link",
-                  url,
-                  children: [
-                    {
-                      type: "text",
-                      value: alias ?? fp,
-                    },
-                  ],
-                }
-              },
-            ])
-          }
-
-          if (opts.highlight) {
-            replacements.push([
-              highlightRegex,
-              (_value: string, ...capture: string[]) => {
-                const [inner] = capture
-                return {
-                  type: "html",
-                  value: `<span class="text-highlight">${inner}</span>`,
-                }
-              },
-            ])
-          }
-
-          if (opts.parseArrows) {
-            replacements.push([
-              arrowRegex,
-              (value: string, ..._capture: string[]) => {
-                const maybeArrow = arrowMapping[value]
-                if (maybeArrow === undefined) return SKIP
-                return {
-                  type: "html",
-                  value: `<span>${maybeArrow}</span>`,
-                }
-              },
-            ])
-          }
-
-          if (opts.parseTags) {
-            replacements.push([
-              tagRegex,
-              (_value: string, tag: string) => {
-                // Check if the tag only includes numbers
-                if (/^\d+$/.test(tag)) {
-                  return false
-                }
-
-                tag = slugTag(tag)
-                if (file.data.frontmatter) {
-                  const noteTags = file.data.frontmatter.tags ?? []
-                  file.data.frontmatter.tags = [...new Set([...noteTags, tag])]
-                }
-
-                return {
-                  type: "link",
-                  url: base + `/tags/${tag}`,
-                  data: {
-                    hProperties: {
-                      className: ["tag-link"],
-                    },
-                  },
-                  children: [
-                    {
-                      type: "text",
-                      value: tag,
-                    },
-                  ],
-                }
-              },
-            ])
-          }
-
-          if (opts.enableInHtmlEmbed) {
-            visit(tree, "html", (node: Html) => {
-              for (const [regex, replace] of replacements) {
-                if (typeof replace === "string") {
-                  node.value = node.value.replace(regex, replace)
-                } else {
-                  node.value = node.value.replace(regex, (substring: string, ...args) => {
-                    const replaceValue = replace(substring, ...args)
-                    if (typeof replaceValue === "string") {
-                      return replaceValue
-                    } else if (Array.isArray(replaceValue)) {
-                      return replaceValue.map(mdastToHtml).join("")
-                    } else if (typeof replaceValue === "object" && replaceValue !== null) {
-                      return mdastToHtml(replaceValue)
-                    } else {
-                      return substring
-                    }
-                  })
-                }
-              }
-            })
-          }
           mdastFindReplace(tree, replacements)
         }
       })
 
-      if (opts.enableVideoEmbed) {
-        plugins.push(() => {
-          return (tree: Root, _file) => {
-            visit(tree, "image", (node, index, parent) => {
-              if (parent && index != undefined && videoExtensionRegex.test(node.url)) {
-                const newNode: Html = {
-                  type: "html",
-                  value: `<video controls src="${node.url}"></video>`,
-                }
-
-                parent.children.splice(index, 1, newNode)
-                return SKIP
-              }
-            })
-          }
-        })
-      }
-
-      if (opts.callouts) {
-        plugins.push(() => {
-          return (tree: Root, _file) => {
-            visit(tree, "blockquote", (node) => {
-              if (node.children.length === 0) {
-                return
-              }
-
-              // find first line
-              const firstChild = node.children[0]
-              if (firstChild.type !== "paragraph" || firstChild.children[0]?.type !== "text") {
-                return
-              }
-
-              const text = firstChild.children[0].value
-              const restOfTitle = firstChild.children.slice(1)
-              const [firstLine, ...remainingLines] = text.split("\n")
-              const remainingText = remainingLines.join("\n")
-
-              const match = firstLine.match(calloutRegex)
-              if (match && match.input) {
-                const [calloutDirective, typeString, collapseChar] = match
-                const calloutType = canonicalizeCallout(typeString.toLowerCase())
-                const collapse = collapseChar === "+" || collapseChar === "-"
-                const defaultState = collapseChar === "-" ? "collapsed" : "expanded"
-                const titleContent = match.input.slice(calloutDirective.length).trim()
-                const useDefaultTitle = titleContent === "" && restOfTitle.length === 0
-                const titleNode: Paragraph = {
-                  type: "paragraph",
-                  children: [
-                    {
-                      type: "text",
-                      value: useDefaultTitle ? capitalize(typeString) : titleContent + " ",
-                    },
-                    ...restOfTitle,
-                  ],
-                }
-                const title = mdastToHtml(titleNode)
-
-                const toggleIcon = `<div class="fold-callout-icon"></div>`
-
-                const titleHtml: Html = {
-                  type: "html",
-                  value: `<div
-                  class="callout-title"
-                >
-                  <div class="callout-icon"></div>
-                  <div class="callout-title-inner">${title}</div>
-                  ${collapse ? toggleIcon : ""}
-                </div>`,
-                }
-
-                const blockquoteContent: (BlockContent | DefinitionContent)[] = [titleHtml]
-                if (remainingText.length > 0) {
-                  blockquoteContent.push({
-                    type: "paragraph",
-                    children: [
-                      {
-                        type: "text",
-                        value: remainingText,
-                      },
-                    ],
-                  })
-                }
-
-                // replace first line of blockquote with title and rest of the paragraph text
-                node.children.splice(0, 1, ...blockquoteContent)
-
-                const classNames = ["callout", calloutType]
-                if (collapse) {
-                  classNames.push("is-collapsible")
-                }
-                if (defaultState === "collapsed") {
-                  classNames.push("is-collapsed")
-                }
-
-                // add properties to base blockquote
-                node.data = {
-                  hProperties: {
-                    ...(node.data?.hProperties ?? {}),
-                    className: classNames.join(" "),
-                    "data-callout": calloutType,
-                    "data-callout-fold": collapse,
-                  },
-                }
-              }
-            })
-          }
-        })
-      }
-
-      if (opts.mermaid) {
-        plugins.push(() => {
-          return (tree: Root, _file) => {
-            visit(tree, "code", (node: Code) => {
-              if (node.lang === "mermaid") {
-                node.data = {
-                  hProperties: {
-                    className: ["mermaid"],
-                  },
-                }
-              }
-            })
-          }
-        })
-      }
-
-      // TODO: add option for task status
-      if (true) {
-        plugins.push(() => {
-          return (tree: Root, _file) => {
-            visit(tree, "listItem", (node) => {
-              // we either have a GFM type checkbox (with a space or x), or a normal list item
-              if (typeof node.checked === "boolean") {
-                node.data = {
-                  ...node.data,
-                  hProperties: {
-                    dataTask: node.checked ? "x" : " ",
-                  },
-                }
-              } else {
-                const paragraph = node.children[0] as Paragraph;
-                if (paragraph?.type !== "paragraph") return;
-
-                const text = paragraph.children?.[0];
-                if (text.type !== 'text') return;
-
-                extendedTaskItemRegex.lastIndex = 0;
-                const result = extendedTaskItemRegex.exec(text.value);
-
-                if (result) {
-                  node.checked = true;
-                  node.data = {
-                    ...node.data,
-                    hProperties: {
-                      dataTask: result[1],
-                    },
-                  };
-                  text.value = result[2];
-                }
-              }
-            });
-          };
-        })
-      }
+      if (opts.enableVideoEmbed) plugins.push(videoEmbedPlugin)
+      if (opts.callouts) plugins.push(calloutPlugin)
+      if (opts.mermaid) plugins.push(mermaidPlugin)
+      if (opts.customTaskStatuses) plugins.push(customTaskStatusPlugin)
 
       return plugins;
     },
@@ -527,186 +96,19 @@ export const ObsidianFlavoredMarkdown: QuartzTransformerPlugin<Partial<Options> 
     htmlPlugins() {
       const plugins: PluggableList = [rehypeRaw]
 
-      if (opts.parseBlockReferences) {
-        plugins.push(() => {
-          const inlineTagTypes = new Set(["p", "li"])
-          const blockTagTypes = new Set(["blockquote"])
-          return (tree: HtmlRoot, file) => {
-            file.data.blocks = {}
-
-            visit(tree, "element", (node, index, parent) => {
-              if (blockTagTypes.has(node.tagName)) {
-                const nextChild = parent?.children.at(index! + 2) as Element
-                if (nextChild && nextChild.tagName === "p") {
-                  const text = nextChild.children.at(0) as Literal
-                  if (text && text.value && text.type === "text") {
-                    const matches = text.value.match(blockReferenceRegex)
-                    if (matches && matches.length >= 1) {
-                      parent!.children.splice(index! + 2, 1)
-                      const block = matches[0].slice(1)
-
-                      if (!Object.keys(file.data.blocks!).includes(block)) {
-                        node.properties = {
-                          ...node.properties,
-                          id: block,
-                        }
-                        file.data.blocks![block] = node
-                      }
-                    }
-                  }
-                }
-              } else if (inlineTagTypes.has(node.tagName)) {
-                const last = node.children.at(-1) as Literal
-                if (last && last.value && typeof last.value === "string") {
-                  const matches = last.value.match(blockReferenceRegex)
-                  if (matches && matches.length >= 1) {
-                    last.value = last.value.slice(0, -matches[0].length)
-                    const block = matches[0].slice(1)
-
-                    if (last.value === "") {
-                      // this is an inline block ref but the actual block
-                      // is the previous element above it
-                      let idx = (index ?? 1) - 1
-                      while (idx >= 0) {
-                        const element = parent?.children.at(idx)
-                        if (!element) break
-                        if (element.type !== "element") {
-                          idx -= 1
-                        } else {
-                          if (!Object.keys(file.data.blocks!).includes(block)) {
-                            element.properties = {
-                              ...element.properties,
-                              id: block,
-                            }
-                            file.data.blocks![block] = element
-                          }
-                          return
-                        }
-                      }
-                    } else {
-                      // normal paragraph transclude
-                      if (!Object.keys(file.data.blocks!).includes(block)) {
-                        node.properties = {
-                          ...node.properties,
-                          id: block,
-                        }
-                        file.data.blocks![block] = node
-                      }
-                    }
-                  }
-                }
-              }
-            })
-
-            file.data.htmlAst = tree
-          }
-        })
-      }
-
-      if (opts.enableYouTubeEmbed) {
-        plugins.push(() => {
-          return (tree: HtmlRoot) => {
-            visit(tree, "element", (node) => {
-              if (node.tagName === "img" && typeof node.properties.src === "string") {
-                const match = node.properties.src.match(ytLinkRegex)
-                const videoId = match && match[2].length == 11 ? match[2] : null
-                const playlistId = node.properties.src.match(ytPlaylistLinkRegex)?.[1]
-                if (videoId) {
-                  // YouTube video (with optional playlist)
-                  node.tagName = "iframe"
-                  node.properties = {
-                    class: "external-embed",
-                    allow: "fullscreen",
-                    frameborder: 0,
-                    width: "600px",
-                    height: "350px",
-                    src: playlistId
-                      ? `https://www.youtube.com/embed/${videoId}?list=${playlistId}`
-                      : `https://www.youtube.com/embed/${videoId}`,
-                  }
-                } else if (playlistId) {
-                  // YouTube playlist only.
-                  node.tagName = "iframe"
-                  node.properties = {
-                    class: "external-embed",
-                    allow: "fullscreen",
-                    frameborder: 0,
-                    width: "600px",
-                    height: "350px",
-                    src: `https://www.youtube.com/embed/videoseries?list=${playlistId}`,
-                  }
-                }
-              }
-            })
-          }
-        })
-      }
-
-      if (opts.enableCheckbox) {
-        plugins.push(() => {
-          return (tree: HtmlRoot, _file) => {
-            visit(tree, "element", (node) => {
-              if (node.tagName === "input" && node.properties.type === "checkbox") {
-                const isChecked = node.properties?.checked ?? false
-                node.properties = {
-                  type: "checkbox",
-                  disabled: false,
-                  checked: isChecked,
-                  class: "checkbox-toggle",
-                }
-              }
-            })
-          }
-        })
-      }
+      if (opts.parseBlockReferences) plugins.push(blockReferencePlugin)
+      if (opts.enableYouTubeEmbed) plugins.push(embedYoutubePlugin)
+      if (opts.enableCheckbox) plugins.push(checkboxPlugin);
 
       return plugins
     },
+
     externalResources() {
       const js: JSResource[] = []
 
-      if (opts.enableCheckbox) {
-        js.push({
-          script: checkboxScript,
-          loadTime: "afterDOMReady",
-          contentType: "inline",
-        })
-      }
-
-      if (opts.callouts) {
-        js.push({
-          script: calloutScript,
-          loadTime: "afterDOMReady",
-          contentType: "inline",
-        })
-      }
-
-      if (opts.mermaid) {
-        js.push({
-          script: `
-          let mermaidImport = undefined
-          document.addEventListener('nav', async () => {
-            if (document.querySelector("code.mermaid")) {
-              mermaidImport ||= await import('https://cdnjs.cloudflare.com/ajax/libs/mermaid/10.7.0/mermaid.esm.min.mjs')
-              const mermaid = mermaidImport.default
-              const darkMode = document.documentElement.getAttribute('saved-theme') === 'dark'
-              mermaid.initialize({
-                startOnLoad: false,
-                securityLevel: 'loose',
-                theme: darkMode ? 'dark' : 'default'
-              })
-
-              await mermaid.run({
-                querySelector: '.mermaid'
-              })
-            }
-          });
-          `,
-          loadTime: "afterDOMReady",
-          moduleType: "module",
-          contentType: "inline",
-        })
-      }
+      if (opts.enableCheckbox) checkboxResources(js)
+      if (opts.callouts) calloutResources(js)
+      if (opts.mermaid) mermaidResources(js)
 
       return { js }
     },
@@ -719,6 +121,7 @@ declare module "vfile" {
     htmlAst: HtmlRoot
   }
 }
+
 declare module "mdast" {
   interface ListItem {
     marker: string
